new progress bar component that all the button components inherit from
