+++ qcp/gui/components/__init__.py
@@ -20,5 +20,6 @@ from qcp.gui.components.abstract_component import \
 from qcp.gui.components.combo_box_component import \
     ComboBoxComponent  # noqa: F401
 from qcp.gui.components.graph_component import GraphComponent  # noqa: F401
+from qcp.gui.components.progress_bar_component import ProgressBarComponent  # noqa: F401
 from qcp.gui.components.threaded_compute import \
     SimulateAlgorithmThread  # noqa: F401
