+++ qcp/gui/components/grovers/button_component.py
@@ -15,16 +15,16 @@ import time
 from typing import List
 
 from PySide6 import QtCore, QtWidgets
-from qcp.gui.components import AbstractComponent
-from qcp.gui.components.constants import (BUTTON_PROGRESS_BAR,
-                                          BUTTON_PROGRESS_BAR_TICK_RATE)
+from qcp.gui.components import ProgressBarComponent
+from qcp.gui.components.constants import (PROGRESS_BAR,
+                                          PROGRESS_BAR_TICK_RATE)
 from qcp.gui.components.grovers.constants import (BUTTON_CANCEL_SEARCH_BUTTON,
                                                   BUTTON_SEARCH_BUTTON)
 from qcp.gui.components.grovers.input_component import GroverInputComponent
 from qcp.gui.constants import THREAD_PAUSE
 
 
-class GroverButtonComponent(AbstractComponent):
+class GroverButtonComponent(ProgressBarComponent):
     """
     Component of the UI that handles button click behaviour.
     """
@@ -54,20 +54,14 @@ class GroverButtonComponent(AbstractComponent):
         """
         super().setup_signals()
 
-        self.progress_bar.hide()
-        self.progress_bar.reset()
-
         self.cancel_button.hide()
-
-        self.pb_thread = ProgressBarThread(self.progress_bar.minimum(),
-                                           self.progress_bar.maximum())
-        self.pb_thread.progress_bar_value_change.connect(self._draw_progress)
-        self.pb_thread.finished.connect(self._hide_progress_bar)
 
         self.search_button.clicked.connect(self.initiate_search)
         self.cancel_button.clicked.connect(self.cancel_search)
 
     def _find_widgets(self):
+        super()._find_widgets()
+
         buttons: List[QtWidgets.QPushButton] = \
             self.main_window.ui_component.findChildren(
                 QtWidgets.QPushButton)
@@ -76,12 +70,6 @@ class GroverButtonComponent(AbstractComponent):
                 self.search_button = b
             if b.objectName() == BUTTON_CANCEL_SEARCH_BUTTON:
                 self.cancel_button = b
-
-        progress_bars = self.main_window.ui_component.findChildren(
-            QtWidgets.QProgressBar)
-        for pb in progress_bars:
-            if pb.objectName() == BUTTON_PROGRESS_BAR:
-                self.progress_bar: QtWidgets.QProgressBar = pb
 
     def initiate_search(self):
         """
@@ -121,74 +109,3 @@ class GroverButtonComponent(AbstractComponent):
             time.sleep(THREAD_PAUSE)
 
         self.cancel_button.hide()
-
-    def tick_progress_bar(self):
-        """
-        Set up the QThread to increment the progress bar widget every
-        tick, and reset it when it fills, to visualise that a calculation
-        is being run by the UI.
-        """
-        if not self.pb_thread.isRunning():
-            self.pb_thread.exiting = False
-            self.pb_thread.start()
-            # wait until the thread has started
-            while not self.pb_thread.isRunning():
-                time.sleep(THREAD_PAUSE)
-
-    @QtCore.Slot(int)
-    def _draw_progress(self, val: int):
-        """
-        Set the progress bar widget to the provided value whenever the
-        signal is called.
-
-        :param int val: The value to set the progress bar to, no checks
-        are performed on it's value.
-        """
-        if self.progress_bar.isHidden():
-            self.progress_bar.show()
-        self.progress_bar.setValue(val)
-
-    def _hide_progress_bar(self):
-        """
-        Hide the progress bar widget.
-        """
-        self.progress_bar.hide()
-
-
-class ProgressBarThread(QtCore.QThread):
-    """
-    Thread that handles visually updating the progress bar value
-    every tick, so that it fills slowly, then resets to zero when
-    full.
-    """
-    progress_bar_value_change = QtCore.Signal(int)
-
-    def __init__(self, min: int, max: int, parent=None):
-        """
-        Setup the ProgressBarThread QThread, to calculate the progress
-        bar ticker value every iteration.
-
-        :param int min: The minimum value allowed for the progress bar
-        :param int max: The maximum allowed value for the progress bar
-        """
-        super().__init__(parent)
-        self.exiting = False
-        self.min = min
-        self.max = max
-        self.pb_val = 0
-
-    def run(self):
-        """
-        Startup the thread, and run the desired actions every loop.
-        """
-        self.pb_val = 0
-        while not self.exiting:
-            val = (self.pb_val + 1) % self.max
-            if val < self.min:
-                val = self.min
-
-            self.progress_bar_value_change.emit(val)
-            self.pb_val = val
-            self.msleep(BUTTON_PROGRESS_BAR_TICK_RATE)
-
-        self.quit()
